discard 'min_shell_separation' shell from available_site_indices
In the current implementation of random doping distribution,
non-interaction or overlap of doping regions is avoided inherently
within the code by discarding num_shells*2 from available_site_indices
despite providing min_shell_separation as an input parameter, which is
not being used.

Changes made in this commit ensure min_shell_separation as the deciding
parameter in the generating the doping distribution.